Validation remove from Labels, Options and Reference

+++ API/utilis.py
@@ -4,10 +4,12 @@ def delete_cases(session):
             case_categories_types = case.category_type.all()
             for category_type in case_categories_types:
                 category_type.options.all().delete()
-                category_type.image.all().delete()
+                if category_type.image.all():
+                    category_type.image.all().delete()
             case.category_type.all().delete()
             case.labels.all().delete()
-            case.reference_folder.image.all().delete()
-            case.reference_folder.delete()
+            if case.reference_folder:                 
+                 case.reference_folder.image.all().delete()
+                 case.reference_folder.delete()
         session.case.all().delete()
         session.slice.all().delete()
